refactor(fetcher): rename project_name/project → git_repo_name/github_repo
The `project` prefix didn't reflect what these fetchers actually do:
`project_name` reads the git remote (gix), and `project` hits the GitHub
REST API. Move them into the matching subdirectories and rename to
`git_repo_name` / `github_repo` so the names match their namespace.

- `src/fetcher/git/repo_name.rs` (GitRepoName)
- `src/fetcher/github/repo.rs` (GithubRepo)
- Template `project_github_activity.toml` and the `derive_shape` test
  follow the new names.

Refs #81.
